+++ engine/musicbrainz.rb
@@ -87,6 +87,33 @@ module Onebox
         end
       end
 
+      def caa_rg_image(rgid)
+        return nil if !SiteSetting.musicbrainz_load_caa_images
+        coverart_url = nil
+        begin
+          api_url = "https://coverartarchive.org/release-group/#{rgid}"
+          result = request_json(api_url)
+
+          image = result["images"][0] if result["images"]
+          if image && image["thumbnails"]
+            thumbnails = image["thumbnails"]
+            @data[:image] = thumbnails["500"] ||
+              thumbnails["large"] ||
+              thumbnails["250"] ||
+              thumbnails["small"]
+            @data[:image_source] = result["release"]
+            @data[:image_source_label] = "Cover Art Archive"
+          end
+
+        rescue OpenURI::HTTPError => e
+          # 404 means the release group does not exist or has no cover art.
+          # Everything else is unexpected and logged as an error.
+          Rails.logger.error e.message unless e.io.status[0] == "404"
+        end
+
+        return coverart_url
+      end
+
       def wikidata(entity=nil)
         return nil if !wikidata_allowed?
         wikidata_rel = get_relations("url", ["wikidata"], entity: entity).first